somewhat fix the resolver and some closure issues
this doesn't completely fix everything, but it's a start.

for the resolver, i deviated from the book and am using Map[Int, Int] to
store the resolved scope depths, instead of Map[Expr, Int].  the reason
is because my Expr leaf classes are all case classes, and scala
auto-generates hashCode() methods for them, which can easily make some
Expr instances hash to the same value (because they're identical, even
though they're different instances).  the correct operation of the
resolver requires that the scope offset storage treat each Expr as if
it's unique.  so, instead, we use System.identityHashCode(expr) as the
key for the map.

next we have closure stuff.  there were two issues here, that are unique
to my decision to have immutable Environment instances:

1) there's a chicken-and-egg problem when defining a function
declaration.  we need to pass the current environment to the function
instance when we create it, but then after we create it, we define it in
the environment.  we want that definition in the environment passed to
the function instance, but we've already passed the old, immutable
environment instance that doesn't have the definition.  so, inside
LoxFunction we now hack the definition into the closure environment.
this hack makes recursive functions work.

2) again, because of our immutable environments, we have a problem with
the parent-scoped environments in the closure environment not matching
up with the environment we had going into the function.  this means:

  a. if an outer-scoped variable has changed value since the function
  was defined, the function itself will see the older value.

  b. if the function itself modifies variables outside its scope, then
  we lose those changes after the function returns.

to combat this, when we enter a function, we first "merge" the caller's
environment into the closure environment, and then when the function
exits, we merge the function's environment (minus the function's actual
scope) back into the caller's environment.  we do this merge on each
active scope, matching by the environment/scope's ID to see which
environments match up.

while this is turning out to be a lot harder than the java version of
the lox interpreter, i'm finding it worthwhile to work through these
sorts of issues.