hopefully fix all remaining issues
last was a problem with Environment.assignAt() and assignRoot(): i was
forgetting that those will return the environment at the point of
assignment, effectively discarding all higher/inner scopes.  the fix is
to chop off the outer environments starting with the assign environment,
tand then take the returned environment of assignHere() and append that.