implement 'continue' in for loops
this required (sadly) creating a new statement type for a for loop.
previously it shared the type for while loops, but unfortunately to
implement 'continue' we need access to the increment expression from the
for loop call, but to re-use the while statement type, we stuff it in an
enclosing block where we can't get to it anymore.